Abstract `DomainBundleProposer` and integrate it into {system, core}_bundle_producer
