+++ domains/client/domain-executor/src/lib.rs
@@ -68,6 +68,7 @@ mod core_domain_worker;
 mod core_executor;
 mod domain_block_processor;
 mod domain_bundle_producer;
+mod domain_bundle_proposer;
 mod fraud_proof;
 mod merkle_tree;
 mod system_bundle_processor;
