+++ domains/client/domain-executor/src/system_bundle_producer.rs
@@ -1,27 +1,23 @@
 use crate::bundle_election_solver::BundleElectionSolver;
 use crate::domain_bundle_producer::ReceiptInterface;
+use crate::domain_bundle_proposer::DomainBundleProposer;
 use crate::utils::ExecutorSlotInfo;
-use crate::{BundleSender, ExecutionReceiptFor};
-use codec::{Decode, Encode};
+use crate::BundleSender;
+use codec::Decode;
 use domain_runtime_primitives::{AccountId, DomainCoreApi};
-use futures::{select, FutureExt};
 use sc_client_api::{AuxStore, BlockBackend, ProofProvider};
-use sc_transaction_pool_api::InPoolTransaction;
 use sp_api::{NumberFor, ProvideRuntimeApi};
 use sp_block_builder::BlockBuilder;
 use sp_blockchain::HeaderBackend;
-use sp_consensus_slots::Slot;
 use sp_domains::{
-    Bundle, BundleHeader, DomainId, ExecutorApi, ExecutorPublicKey, ExecutorSignature,
-    SignedBundle, SignedOpaqueBundle,
+    DomainId, ExecutorApi, ExecutorPublicKey, ExecutorSignature, SignedBundle, SignedOpaqueBundle,
 };
 use sp_keystore::{SyncCryptoStore, SyncCryptoStorePtr};
 use sp_runtime::generic::BlockId;
-use sp_runtime::traits::{BlakeTwo256, Block as BlockT, Hash as HashT, Zero};
+use sp_runtime::traits::Block as BlockT;
 use sp_runtime::RuntimeAppPublic;
 use std::marker::PhantomData;
 use std::sync::Arc;
-use std::time;
 use subspace_core_primitives::BlockNumber;
 use system_runtime_primitives::SystemDomainApi;
 
@@ -35,11 +31,11 @@ where
     domain_id: DomainId,
     primary_chain_client: Arc<PClient>,
     client: Arc<Client>,
-    transaction_pool: Arc<TransactionPool>,
     bundle_sender: Arc<BundleSender<Block, PBlock>>,
     is_authority: bool,
     keystore: SyncCryptoStorePtr,
     bundle_election_solver: BundleElectionSolver<Block, PBlock, Client>,
+    domain_bundle_proposer: DomainBundleProposer<Block, Client, TransactionPool>,
     _phantom_data: PhantomData<PBlock>,
 }
 
@@ -54,11 +50,11 @@ where
             domain_id: self.domain_id,
             primary_chain_client: self.primary_chain_client.clone(),
             client: self.client.clone(),
-            transaction_pool: self.transaction_pool.clone(),
             bundle_sender: self.bundle_sender.clone(),
             is_authority: self.is_authority,
             keystore: self.keystore.clone(),
             bundle_election_solver: self.bundle_election_solver.clone(),
+            domain_bundle_proposer: self.domain_bundle_proposer.clone(),
             _phantom_data: self._phantom_data,
         }
     }
@@ -138,15 +134,16 @@ where
         keystore: SyncCryptoStorePtr,
     ) -> Self {
         let bundle_election_solver = BundleElectionSolver::new(client.clone(), keystore.clone());
+        let domain_bundle_proposer = DomainBundleProposer::new(client.clone(), transaction_pool);
         Self {
             domain_id,
             primary_chain_client,
             client,
-            transaction_pool,
             bundle_sender,
             is_authority,
             keystore,
             bundle_election_solver,
+            domain_bundle_proposer,
             _phantom_data: PhantomData::default(),
         }
     }
@@ -183,7 +180,13 @@ where
             tracing::info!(target: LOG_TARGET, "📦 Claimed bundle at slot {slot}");
 
             let bundle = self
-                .propose_bundle_at(slot, primary_info, receipt_interface)
+                .domain_bundle_proposer
+                .propose_bundle_at::<PBlock, _, _>(
+                    slot,
+                    primary_info,
+                    receipt_interface,
+                    primary_info.0,
+                )
                 .await?;
 
             let to_sign = bundle.hash();
@@ -224,137 +227,5 @@ where
         } else {
             Ok(None)
         }
-    }
-
-    async fn propose_bundle_at<R>(
-        &self,
-        slot: Slot,
-        primary_info: (PBlock::Hash, NumberFor<PBlock>),
-        receipt_interface: R,
-    ) -> sp_blockchain::Result<Bundle<Block::Extrinsic, NumberFor<PBlock>, PBlock::Hash, Block::Hash>>
-    where
-        R: ReceiptInterface<PBlock::Hash>,
-    {
-        let parent_number = self.client.info().best_number;
-
-        let mut t1 = self.transaction_pool.ready_at(parent_number).fuse();
-        // TODO: proper timeout
-        let mut t2 = futures_timer::Delay::new(time::Duration::from_micros(100)).fuse();
-
-        let pending_iterator = select! {
-            res = t1 => res,
-            _ = t2 => {
-                tracing::warn!(
-                    target: LOG_TARGET,
-                    "Timeout fired waiting for transaction pool at #{}, proceeding with production.",
-                    parent_number,
-                );
-                self.transaction_pool.ready()
-            }
-        };
-
-        // TODO: proper deadline
-        let pushing_duration = time::Duration::from_micros(500);
-
-        let start = time::Instant::now();
-
-        // TODO: Select transactions properly from the transaction pool
-        //
-        // Selection policy:
-        // - minimize the transaction equivocation.
-        // - maximize the executor computation power.
-        let mut extrinsics = Vec::new();
-
-        for pending_tx in pending_iterator {
-            if start.elapsed() >= pushing_duration {
-                break;
-            }
-            let pending_tx_data = pending_tx.data().clone();
-            extrinsics.push(pending_tx_data);
-        }
-
-        let extrinsics_root = BlakeTwo256::ordered_trie_root(
-            extrinsics.iter().map(|xt| xt.encode()).collect(),
-            sp_core::storage::StateVersion::V1,
-        );
-
-        let (primary_hash, primary_number) = primary_info;
-
-        let receipts = if primary_number.is_zero() {
-            Vec::new()
-        } else {
-            self.collect_system_bundle_receipts(primary_hash, parent_number, receipt_interface)?
-        };
-
-        let bundle = Bundle {
-            header: BundleHeader {
-                primary_hash,
-                slot_number: slot.into(),
-                extrinsics_root,
-            },
-            receipts,
-            extrinsics,
-        };
-
-        Ok(bundle)
-    }
-
-    /// Returns the receipts in the next system domain bundle.
-    fn collect_system_bundle_receipts<R>(
-        &self,
-        primary_hash: PBlock::Hash,
-        header_number: NumberFor<Block>,
-        receipt_interface: R,
-    ) -> sp_blockchain::Result<Vec<ExecutionReceiptFor<PBlock, Block::Hash>>>
-    where
-        R: ReceiptInterface<PBlock::Hash>,
-    {
-        let best_execution_chain_number =
-            receipt_interface.best_execution_chain_number(primary_hash)?;
-        let max_drift = receipt_interface.maximum_receipt_drift(primary_hash)?;
-
-        let load_receipt = |block_hash| {
-            crate::aux_schema::load_execution_receipt::<
-                _,
-                Block::Hash,
-                NumberFor<PBlock>,
-                PBlock::Hash,
-            >(&*self.client, block_hash)?
-            .ok_or_else(|| {
-                sp_blockchain::Error::Backend(format!("Receipt not found for {block_hash}"))
-            })
-        };
-
-        let header_number: BlockNumber = header_number
-            .try_into()
-            .unwrap_or_else(|_| panic!("Secondary number must fit into u32; qed"));
-
-        // Ideally, the receipt of current block will be included in the next block, i.e., no
-        // missing receipts.
-        let receipts = if header_number == best_execution_chain_number + 1 {
-            let block_hash = self.client.hash(header_number.into())?.ok_or_else(|| {
-                sp_blockchain::Error::Backend(format!(
-                    "Hash for Block {:?} not found",
-                    header_number
-                ))
-            })?;
-            vec![load_receipt(block_hash)?]
-        } else {
-            // Receipts for some previous blocks are missing.
-            let max_allowed = (best_execution_chain_number + max_drift).min(header_number);
-
-            let mut to_send = best_execution_chain_number + 1;
-            let mut receipts = Vec::with_capacity((max_allowed - to_send + 1) as usize);
-            while to_send <= max_allowed {
-                let block_hash = self.client.hash(to_send.into())?.ok_or_else(|| {
-                    sp_blockchain::Error::Backend(format!("Hash for Block {:?} not found", to_send))
-                })?;
-                receipts.push(load_receipt(block_hash)?);
-                to_send += 1;
-            }
-            receipts
-        };
-
-        Ok(receipts)
     }
 }
